requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
passlib==1.7.4
ping3==4.0.8
propcache==0.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
python-jose==3.3.0
python-multipart==0.0.18
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
yarl==1.18.0
